requirements-lock.txt
# Production-locked requirements for Vispio AI Image Captioning
# Generated from working environment - use for production deployment

# Core Streamlit Application
streamlit==1.45.1
altair==5.5.0
blinker==1.8.2
cachetools==5.5.2
click==8.1.7
gitpython==3.1.44
markdown-it-py==3.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==5.29.4
pyarrow==19.0.1
pydeck==0.9.1
python-dateutil==2.9.0.post0
requests==2.31.0
rich==13.7.1
tenacity==8.4.2
toml==0.10.2
tornado==6.4
typing-extensions==4.13.2
tzlocal==5.3.1
validators==0.22.0
watchdog==6.0.0

# Google Gemini AI API
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-auth==2.40.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
proto-plus==1.26.1

# Text-to-Speech and Audio Processing
gtts==2.5.4
pydub==0.25.1
pygame==2.6.0.dev2

# Environment Variables Management
python-dotenv==1.0.1

# HTTP and Network
urllib3==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7

# Core Python Dependencies
setuptools==70.1.0
wheel==0.43.0
six==1.16.0